Fix naming: CatsTraverseWithCatsSpec => CatsTraverseWithoutCatsSpec in the orphan-cats-test-without-cats module
